Mask correctly in Dyadic.mk (#9)
for keys longer than 32 bytes, the `Dyadic.mk` funcion calculated the
mask wrongly, leading to an overflow trap.

The fix is to shift the mask only by `i % 8`.

On shorter prefixes the impact on this bug is that they are not properly
normalized. This does not affect correctness, only that the
representation isn’t unique.

Fixes #8.